V0 du modele
