--- smart_doodle_back/src/main/java/org/esir/smartdoodle/domain/Event.java
@@ -1,4 +1,0 @@
-package org.esir.smartdoodle.domain;
-
-public class Event {
-}
