+++ smart_doodle_back/src/main/java/org/esir/smartdoodle/domain/PollService.java
@@ -1,23 +1,26 @@
 package org.esir.smartdoodle.domain;
 
+import java.util.List;
+
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
-import java.util.List;
+
+import org.esir.smartdoodle.entities.Poll;
 
 @ApplicationScoped
-public class GiftService {
+public class PollService {
 
     @Inject
     private EntityManager em;
 
     @Transactional
-    public void set(Gift gift) {
-        em.persist(gift);
+    public void set(Poll poll) {
+        em.persist(poll);
     }
 
-    public List<Gift> list() {
-        return em.createQuery("from Gift").getResultList();
+    public List<Poll> list() {
+        return em.createQuery("from Poll").getResultList();
     }
 }
