+++ smart_doodle_back/src/main/java/org/esir/smartdoodle/entities/Answer.java
@@ -1,4 +1,4 @@
-package entities;
+package org.esir.smartdoodle.entities;
 
 /**
  * Created by 16002492 on 24/04/19.
