+++ smart_doodle_back/src/main/java/org/esir/smartdoodle/entities/Slot.java
@@ -1,4 +1,4 @@
-package entities;
+package org.esir.smartdoodle.entities;
 
 import javax.persistence.*;
 import java.util.Date;
