+++ smart_doodle_back/src/main/java/org/esir/smartdoodle/jpa/EntityManagerHelper.java
@@ -1,4 +1,4 @@
-package jpa;
+package org.esir.smartdoodle.jpa;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
